Add Documentation, Refactor
